splitted the ui in a subdir
